refactor: converter 패키지 생성
변경된 점:
- Jpa의 배열 저장을 위한 Converter 클래스를 하나의 패키지 아래에 저장
