+++ src/main/java/com/mtvs/devlinkbackend/util/converter/IntegerListConverter.java
@@ -1,4 +1,4 @@
-package com.mtvs.devlinkbackend.util;
+package com.mtvs.devlinkbackend.util.converter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -10,12 +10,12 @@ import java.util.Arrays;
 import java.util.List;
 
 @Converter
-public class LongListConverter implements AttributeConverter<List<Long>, String> {
+public class IntegerListConverter implements AttributeConverter<List<Integer>, String> {
 
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     @Override
-    public String convertToDatabaseColumn(List<Long> attribute) {
+    public String convertToDatabaseColumn(List<Integer> attribute) {
         try {
             return objectMapper.writeValueAsString(attribute);
         } catch (JsonProcessingException e) {
@@ -24,9 +24,9 @@ public class LongListConverter implements AttributeConverter<List<Long>, String>
     }
 
     @Override
-    public List<Long> convertToEntityAttribute(String dbData) {
+    public List<Integer> convertToEntityAttribute(String dbData) {
         try {
-            return Arrays.asList(objectMapper.readValue(dbData, Long[].class));
+            return Arrays.asList(objectMapper.readValue(dbData, Integer[].class));
         } catch (IOException e) {
             throw new RuntimeException("Could not convert JSON to list", e);
         }
